+++ models/MaterialConsumeModel.js
@@ -16,7 +16,7 @@ const MaterialConsume = sequelize.define(
             compositePrimaryKey: true,
         },
         units: {
-            type: DataTypes.DOUBLE(12, 2),
+            type: DataTypes.FLOAT,
             allowNull: false,
         },
     },
